+++ front/components/home/content/Solutions/BenefitsSection.tsx
