+++ front/components/home/content/Solutions/CustomerStoriesSection.tsx
