+++ front/components/home/content/Solutions/HeroSection.tsx
