+++ front/components/home/content/Solutions/UseCasesSection.tsx
